refactor:createMap 분리
